fix: update to use renamed redis_utils
